restService/requirements.txt
requests
flask
pytest
pyHamcrest